Fix tying error.

+++ src/MagnusAddon/anki_extentions/deck_configdict_ex.py
@@ -1,9 +1,9 @@
 from anki.decks import DeckConfigDict
-from sysutils.typed import int_
+from sysutils.typed import *
 
 class DeckConfigDictEx:
     def __init__(self, deck_config_dict:DeckConfigDict) -> None:
         self._dict = deck_config_dict
 
-    def get_seconds_to_show_question(self) -> int:
-        return int_(self._dict.get('secondsToShowQuestion'))
+    def get_seconds_to_show_question(self) -> float:
+        return float_(self._dict.get('secondsToShowQuestion'))
